add unreviewed appointments
